Patch google.protobuf.Value and speed up crd generation
Cue lang protobuf decoder builds`google.protobuf.Value` as Top (_),
meaning it can be any type.
But OpenAPI spec does not allow this generic type and requires something
more concrete.

To let build crds that pass the k8s validation, this commit makes the
google.protobuf.Value to end with a `type: object` in the crd's OpenAPI spec.

This is not the final solution since the affected field will not accept
any kind of type, but only structs.